git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@6808 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/DemoFile.cpp
@@ -134,7 +134,7 @@ idDemoFile::Log
 */
 void idDemoFile::Log(const char *p) {
 	if ( fLog && p && *p ) {
-		fLog->Write( p, strlen(p) );
+        fLog->Write(p, static_cast<int>(strlen(p)));
 	}
 }
 
@@ -207,7 +207,7 @@ const char *idDemoFile::ReadHashString() {
 
 	if ( log && fLog ) {
 		const char *text = va( "%s > Reading hash string\n", logStr.c_str() );
-		fLog->Write( text, strlen( text ) );
+        fLog->Write(text, static_cast<int>(strlen(text)));
 	} 
 
 	ReadInt( index );
@@ -241,7 +241,7 @@ idDemoFile::WriteHashString
 void idDemoFile::WriteHashString( const char *str ) {
 	if ( log && fLog ) {
 		const char *text = va( "%s > Writing hash string\n", logStr.c_str() );
-		fLog->Write( text, strlen( text ) );
+        fLog->Write(text, static_cast<int>(strlen(text)));
 	}
 	// see if it is already in the has table
 	for ( int i = 0 ; i < demoStrings.Num() ; i++ ) {
