+++ game/HidingSpotSearchCollection.cpp
@@ -50,7 +50,7 @@ void CHidingSpotSearchCollection::clear()
 
 void CHidingSpotSearchCollection::Save(idSaveGame *savefile) const
 {
-	savefile->WriteInt(searches.size());
+    savefile->WriteInt(static_cast<int>(searches.size()));
 	for (HidingSpotSearchMap::const_iterator i = searches.begin(); i != searches.end(); ++i)
 	{
 		const HidingSpotSearchNodePtr& node = i->second;
