+++ idlib/CmdArgs.cpp
@@ -67,7 +67,7 @@ const char *idCmdArgs::Args(  int start, int end, bool escapeArgs ) const {
 				if ( *p == '\\' ) {
 					strcat( cmd_args, "\\\\" );
 				} else {
-					int l = strlen( cmd_args );
+					int l = static_cast<int>(strlen( cmd_args ));
 					cmd_args[ l ] = *p;
 					cmd_args[ l+1 ] = '\0';
 				}
@@ -106,7 +106,7 @@ void idCmdArgs::TokenizeString( const char *text, bool keepAsStrings ) {
 		return;
 	}
 
-	lex.LoadMemory( text, strlen( text ), "idCmdSystemLocal::TokenizeString" );
+	lex.LoadMemory( text, static_cast<int>(strlen( text )), "idCmdSystemLocal::TokenizeString" );
 	lex.SetFlags( LEXFL_NOERRORS
 				| LEXFL_NOWARNINGS
 				| LEXFL_NOSTRINGCONCAT
@@ -172,7 +172,7 @@ void idCmdArgs::AppendArg( const char *text ) {
 		idStr::Copynz( tokenized, text, sizeof( tokenized ) );
 	} else {
 		argv[ argc ] = argv[ argc-1 ] + strlen( argv[ argc-1 ] ) + 1;
-		idStr::Copynz( argv[ argc ], text, sizeof( tokenized ) - ( argv[ argc ] - tokenized ) );
+        idStr::Copynz(argv[argc], text, static_cast<int>(sizeof(tokenized) - (argv[argc] - tokenized)));
 		argc++;
 	}
 }
