+++ idlib/math/Simd_SSE2.cpp
@@ -67,7 +67,7 @@ void VPCALL idSIMD_SSE2::CmpLT( byte *dst, const byte bitNum, const float *src0,
 	int dst_l;
 	
 	/* if the float array is not aligned on a 4 byte boundary */
-	if ( ((int) src0) & 3 ) {
+	if ( ptrdiff_t(src0) & 3 ) {
 		/* unaligned memory access */
 		pre = 0;
 		cnt = count >> 2;
@@ -143,8 +143,8 @@ void VPCALL idSIMD_SSE2::CmpLT( byte *dst, const byte bitNum, const float *src0,
 	}																					
 	else {																				
 		/* aligned memory access */														
-		aligned = (float *) ((((int) src0) + 15) & ~15);								
-		if ( (int)aligned > ((int)src0) + count ) {										
+		aligned = (float *) ((ptrdiff_t(src0) + 15) & ~15);
+		if ( ptrdiff_t(aligned) > ptrdiff_t(src0) + count ) {
 			pre = count;																
 			post = 0;																	
 		}																				
@@ -235,7 +235,7 @@ void VPCALL idSIMD_SSE2::CmpLT( byte *dst, const byte bitNum, const float *src0,
 	}
 }
 
-#elif defined(_WIN32)
+#elif defined(_MSC_VER) && defined(_M_IX86)
 
 #include <xmmintrin.h>
 
@@ -252,10 +252,10 @@ void VPCALL idSIMD_SSE2::CmpLT( byte *dst, const byte bitNum, const float *src0,
 ALIGN8_INIT1( unsigned short SIMD_W_zero, 0 );
 ALIGN8_INIT1( unsigned short SIMD_W_maxShort, 1<<15 );
 
-ALIGN4_INIT4( unsigned long SIMD_SP_singleSignBitMask, (unsigned long) ( 1 << 31 ), 0, 0, 0 );
-ALIGN4_INIT1( unsigned long SIMD_SP_signBitMask, (unsigned long) ( 1 << 31 ) );
-ALIGN4_INIT1( unsigned long SIMD_SP_absMask, (unsigned long) ~( 1 << 31 ) );
-ALIGN4_INIT1( unsigned long SIMD_SP_infinityMask, (unsigned long) ~( 1 << 23 ) );
+ALIGN4_INIT4( unsigned int SIMD_SP_singleSignBitMask, (unsigned int) ( 1 << 31 ), 0, 0, 0 );
+ALIGN4_INIT1( unsigned int SIMD_SP_signBitMask, (unsigned int) ( 1 << 31 ) );
+ALIGN4_INIT1( unsigned int SIMD_SP_absMask, (unsigned int) ~( 1 << 31 ) );
+ALIGN4_INIT1( unsigned int SIMD_SP_infinityMask, (unsigned int) ~( 1 << 23 ) );
 
 ALIGN4_INIT1( float SIMD_SP_zero, 0.0f );
 ALIGN4_INIT1( float SIMD_SP_one, 1.0f );
